requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-telegram-bot[job-queue]>=20.6
supabase>=2.3.0
google-generativeai>=0.3.0
python-docx>=1.1.0
XlsxWriter>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
